s3/requirements.txt
# requirements.txt para o serviço s3

fastapi
uvicorn
kafka-python
elasticsearch
redis
